backend/fast_api/requirements.txt
fastapi==0.85.0
sqlalchemy
uvicorn==0.18.3
python-decouple==3.6
requests==2.28.1
pipenv==2022.9.8
psycopg2-binary==2.9.3
pymysql==1.0.2
mysql==0.0.3